requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Flask-Limiter==2.4.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
cryptography==3.4.7
gunicorn==20.1.0

